# doc/api.yaml
openapi: 3.0.0
info:
  title: English for kids
  description: RSSchool task
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    Category:
      required:
        - name
      properties:
        name:
          type: string
        words:
          type: array
          items:
            $ref: '#/components/schemas/Word'
    Word:
      required:
        - word
        - translation
        - audioSrc
      properties:
        word:
          type: string
        translation:
          type: string
        audioSrc:
          type: number
        image:
          type: string
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /categories:
    get:
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            description: Determines how many documents should be skipped before retrieval
        - in: query
          name: limit
          schema:
            type: integer
          description: Determines how many documents to retrieve (at most)
      tags:
        - Categories
      summary: Get categories
      description: Returns retrieved
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        400:
          description: Bad request
    post:
      tags:
        - Categories
      summary: Create category
      description: Creates a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  description: Category name
              required:
                - name
      responses:
        200:
          description: The category has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
...